Extract disposables from interactor to plugin
