+++ app-example/build.gradle
@@ -77,6 +77,7 @@ dependencies {
     implementation deps.retrofit.okhttpLoggingInterceptor
 
     implementation project(":libraries:rib-base")
+    implementation project(":libraries:rib-rx")
     implementation project(":libraries:rib-portal")
 
     implementation deps.external.mviCore
