+++ app-example/src/main/java/com/badoo/ribs/example/photo_feed/PhotoFeedBuilder.kt
@@ -3,6 +3,7 @@ package com.badoo.ribs.example.photo_feed
 import com.badoo.ribs.builder.SimpleBuilder
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.example.photo_feed.feature.PhotoFeedFeature
+import com.badoo.ribs.rx.Disposables.Companion.disposeOnDetach
 
 class PhotoFeedBuilder(
     private val dependency: PhotoFeed.Dependency
@@ -19,7 +20,10 @@ class PhotoFeedBuilder(
         return PhotoFeedNode(
             buildParams = buildParams,
             viewFactory = customisation.viewFactory(null),
-            plugins = listOf(interactor)
+            plugins = listOf(
+                interactor,
+                disposeOnDetach(feature)
+            )
         )
     }
