+++ app-example/src/main/java/com/badoo/ribs/example/photo_feed/PhotoFeedInteractor.kt
@@ -15,8 +15,7 @@ internal class PhotoFeedInteractor(
     buildParams: BuildParams<*>,
     private val feature: PhotoFeedFeature
 ) : Interactor<PhotoFeed, PhotoFeedView>(
-    buildParams = buildParams,
-    disposables = feature
+    buildParams = buildParams
 ) {
 
     override fun onViewCreated(view: PhotoFeedView, viewLifecycle: Lifecycle) {
