+++ libraries/rib-base/src/main/java/com/badoo/ribs/clienthelper/interactor/Interactor.kt
@@ -4,19 +4,11 @@ import com.badoo.ribs.android.requestcode.RequestCodeClient
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.Rib
 import com.badoo.ribs.core.modality.BuildParams
-import com.badoo.ribs.core.plugin.BackPressHandler
-import com.badoo.ribs.core.plugin.NodeLifecycleAware
-import com.badoo.ribs.core.plugin.RibAware
-import com.badoo.ribs.core.plugin.RibAwareImpl
-import com.badoo.ribs.core.plugin.SavesInstanceState
-import com.badoo.ribs.core.plugin.SubtreeChangeAware
-import com.badoo.ribs.core.plugin.ViewAware
+import com.badoo.ribs.core.plugin.*
 import com.badoo.ribs.core.view.RibView
-import io.reactivex.disposables.Disposable
 
 abstract class Interactor<R : Rib, V : RibView>(
     private val buildParams: BuildParams<*>,
-    private val disposables: Disposable? = null,
     private val ribAware: RibAware<R> = RibAwareImpl()
 ) : RibAware<R> by ribAware,
     ViewAware<V>,
@@ -28,10 +20,6 @@ abstract class Interactor<R : Rib, V : RibView>(
 
     override val requestCodeClientId: String
         get() = buildParams.identifier.toString()
-
-    override fun onDetach() {
-        disposables?.dispose()
-    }
 
     val node: Node<*>
        get() = rib.node
