+++ libraries/rib-portal-rx/build.gradle
@@ -30,7 +30,7 @@ dependencies {
 
     api project(":libraries:rib-portal")
     implementation project(":libraries:rib-base")
-    implementation project(":libraries:rib-workflows-rx")
+    implementation project(":libraries:rib-rx")
 
     testImplementation deps.android.annotations
     testImplementation deps.test.junit
