+++ libraries/rib-recyclerview/build.gradle
@@ -28,6 +28,7 @@ dependencies {
     implementation deps.external.rxrelay2
     implementation deps.external.rxjava2
     implementation project(":libraries:rib-base")
+    implementation project(":libraries:rib-rx")
 
     testImplementation deps.android.annotations
     testImplementation deps.test.junit
