+++ libraries/rib-recyclerview/src/main/java/com/badoo/ribs/android/recyclerview/RecyclerViewHostInteractor.kt
@@ -12,8 +12,7 @@ internal class RecyclerViewHostInteractor<T : Parcelable>(
     private val feature: RecyclerViewHostFeature<T>,
     private val adapter: Adapter<T>
 ) : Interactor<RecyclerViewHost<T>, RibView>(
-    buildParams = buildParams,
-    disposables = feature
+    buildParams = buildParams
 ) {
 
     override fun onAttach(nodeLifecycle: Lifecycle) {
