+++ libraries/rib-rx/proguard-rules.pro
