+++ settings.gradle
@@ -16,7 +16,7 @@ include ':libraries:rib-debug-utils'
 include ':libraries:rib-portal'
 include ':libraries:rib-portal-rx'
 include ':libraries:rib-recyclerview'
-include ':libraries:rib-workflows-rx'
+include ':libraries:rib-rx'
 include ':sandbox'
 include ':templates:release-0.8'
 include ':templates:release-0.12'
