+++ templates/snapshot/build.gradle
@@ -2,6 +2,7 @@ apply from: file('../common-template.gradle')
 
 dependencies {
     implementation project(":libraries:rib-base")
+    implementation project(":libraries:rib-rx")
     debugImplementation project(":libraries:rib-base-test-activity")
     androidTestImplementation project(":libraries:rib-base-test")
 }
