+++ templates/snapshot/src/main/java/com/badoo/ribs/template/leaf/foo_bar/FooBarNode.kt
@@ -2,12 +2,12 @@ package com.badoo.ribs.template.leaf.foo_bar
 
 import com.badoo.ribs.clienthelper.connector.Connectable
 import com.badoo.ribs.clienthelper.connector.NodeConnector
-import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
 import com.badoo.ribs.core.view.RibView
 import com.badoo.ribs.template.leaf.foo_bar.FooBar.Input
 import com.badoo.ribs.template.leaf.foo_bar.FooBar.Output
+import com.badoo.ribs.workflows.rx.RxWorkflowNode
 import io.reactivex.Single
 
 class FooBarNode internal constructor(
@@ -15,7 +15,7 @@ class FooBarNode internal constructor(
     viewFactory: ((RibView) -> FooBarView?)?,
     plugins: List<Plugin> = emptyList(),
     connector: NodeConnector<Input, Output> = NodeConnector()
-) : Node<FooBarView>(
+) : RxWorkflowNode<FooBarView>(
     buildParams = buildParams,
     viewFactory = viewFactory,
     plugins = plugins
